ProductWeb/requirements.txt
argon2-cffi==18.1.0
arrow==0.12.1
bcrypt==3.1.4
binaryornot==0.4.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cookiecutter==1.6.0
defusedxml==0.5.0
Django==2.0.7
django-bootstrap4==0.0.6
django-braces==1.13.0
django-debug-toolbar==1.9.1
django-filter==1.1.0
djangorestframework==3.8.2
future==0.16.0
idna==2.6
Jinja2==2.10
jinja2-time==0.2.0
Markdown==2.6.11
MarkupSafe==1.0
misaka==2.1.0
mysqlclient==1.3.12
oauthlib==2.1.0
Pillow==5.1.0
poyo==0.4.1
psycopg2==2.7.4
pycparser==2.18
PyJWT==1.6.4
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sqlparse==0.2.4
urllib3==1.22
whichcraft==0.4.1
wincertstore==0.2
